backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
pytest==7.4.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
SQLAlchemy==2.0.36
httpx==0.27.2
